requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
djangorestframework==3.9.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
psycopg2==2.7.7
pytz==2018.9
requests==2.21.0
uritemplate==3.0.0
urllib3==1.24.1
